BlogAPI/requirements.txt
Django==3.2.5
django-rest-framework==3.12.4
django-cors-headers==3.7.0